+++ conf/app.conf.php
@@ -13,22 +13,10 @@ return static function (SeablastConfiguration $SBConfig): void {
     $SBConfig->flag
         //->activate(SeablastConstant::FLAG_WEB_RUNNING)
         ->activate(AuthConstant::FLAG_USE_SOCIAL_LOGIN) // actual social login requires AuthApp:..social.._ID
+        // - AuthApp:GOOGLE_CLIENT_ID
+        // - AuthApp:FACEBOOK_APP_ID
     ;
     $SBConfig
-//        ->setArrayArrayString(// TODO - move to Seablast? or Seablast/Auth?
-//            SeablastConstant::APP_MAPPING,
-//            '/api/facebook-login',
-//            [
-//                'model' => '\Seablast\Auth\Models\ApiFacebookLoginModel',
-//            ]
-//        )
-//        ->setArrayArrayString(// TODO - move to Seablast? or Seablast/Auth?
-//            SeablastConstant::APP_MAPPING,
-//            '/api/google-login',
-//            [
-//                'model' => '\Seablast\Auth\Models\ApiGoogleLoginModel',
-//            ]
-//        )
         // /api/social-login is a single end-point , differentiation by provider is done in the parameter provider;
         // so far just facebook, google
         ->setArrayArrayString(
@@ -42,5 +30,13 @@ return static function (SeablastConfiguration $SBConfig): void {
         //    ->setString(SeablastConstant::SB_PHINX_ENVIRONMENT, 'testing')
         // Expected route for the page where user can log-in/log-out
         ->setString(AuthConstant::USER_ROUTE, '/user')
+        ->setArrayArrayString(
+            SeablastConstant::APP_MAPPING,
+            '/user', // page slug, i.e. URL representation
+            [
+                'template' => 'user', // template used by the View component
+                'model' => '\Seablast\Auth\UserModel',
+            ]
+        )
     ;
 };